Move all dataset files/folders to the new structure
