+++ superset-frontend/src/pages/Dataset/DatasetPage.test.tsx
@@ -17,27 +17,31 @@
  * under the License.
  */
 import React from 'react';
-import fetchMock from 'fetch-mock';
 import { render, screen } from 'spec/helpers/testing-library';
-import EditDataset from './index';
+import DatasetPage from 'src/pages/Dataset/DatasetPage';
 
-const DATASET_ENDPOINT = 'glob:*api/v1/dataset/1/related_objects';
+const mockHistoryPush = jest.fn();
+jest.mock('react-router-dom', () => ({
+  ...jest.requireActual('react-router-dom'),
+  useHistory: () => ({
+    push: mockHistoryPush,
+  }),
+  useParams: () => ({ datasetId: undefined }),
+}));
 
-const mockedProps = {
-  id: '1',
-};
+describe('DatasetPage', () => {
+  it('renders a blank state DatasetPage', async () => {
+    render(<DatasetPage />, { useRedux: true });
 
-fetchMock.get(DATASET_ENDPOINT, { charts: { results: [], count: 2 } });
+    const blankeStateImgs = screen.getAllByRole('img', { name: /empty/i });
 
-test('should render edit dataset view with tabs', async () => {
-  render(<EditDataset {...mockedProps} />);
+    // Header
+    expect(await screen.findByText(/new dataset/i)).toBeVisible();
+    // Left panel
+    expect(blankeStateImgs[0]).toBeVisible();
+    // Footer
+    expect(screen.getByText(/Cancel/i)).toBeVisible();
 
-  const columnTab = await screen.findByRole('tab', { name: /columns/i });
-  const metricsTab = screen.getByRole('tab', { name: /metrics/i });
-  const usageTab = screen.getByRole('tab', { name: /usage/i });
-
-  expect(fetchMock.calls(DATASET_ENDPOINT)).toBeTruthy();
-  expect(columnTab).toBeInTheDocument();
-  expect(metricsTab).toBeInTheDocument();
-  expect(usageTab).toBeInTheDocument();
+    expect(blankeStateImgs.length).toBe(1);
+  });
 });
