+++ superset-frontend/src/pages/DatasetEditPanel/DatasetEditPanel.test.tsx
@@ -16,29 +16,28 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
 import React from 'react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { supersetTheme, ThemeProvider } from '@superset-ui/core';
-import DatasetPanel from './DatasetPanel';
-import { exampleColumns } from './fixtures';
+import fetchMock from 'fetch-mock';
+import { render, screen } from 'spec/helpers/testing-library';
+import DatasetEditPanel from './index';
 
-export default {
-  title: 'Superset App/views/CRUD/data/dataset/DatasetPanel',
-  component: DatasetPanel,
-} as ComponentMeta<typeof DatasetPanel>;
+const DATASET_ENDPOINT = 'glob:*api/v1/dataset/1/related_objects';
 
-export const Basic: ComponentStory<typeof DatasetPanel> = args => (
-  <ThemeProvider theme={supersetTheme}>
-    <div style={{ height: '350px' }}>
-      <DatasetPanel {...args} />
-    </div>
-  </ThemeProvider>
-);
-
-Basic.args = {
-  tableName: 'example_table',
-  loading: false,
-  hasError: false,
-  columnList: exampleColumns,
+const mockedProps = {
+  id: '1',
 };
+
+fetchMock.get(DATASET_ENDPOINT, { charts: { results: [], count: 2 } });
+
+test('should render edit dataset panel with tabs', async () => {
+  render(<DatasetEditPanel {...mockedProps} />);
+
+  const columnTab = await screen.findByRole('tab', { name: /columns/i });
+  const metricsTab = screen.getByRole('tab', { name: /metrics/i });
+  const usageTab = screen.getByRole('tab', { name: /usage/i });
+
+  expect(fetchMock.calls(DATASET_ENDPOINT)).toBeTruthy();
+  expect(columnTab).toBeInTheDocument();
+  expect(metricsTab).toBeInTheDocument();
+  expect(usageTab).toBeInTheDocument();
+});
